@@ -21,32 +21,34 @@
     "fmt": "prettier --write \"**/*.md\""
   },
   "dependencies": {
-    "@docusaurus/core": "2.4.3",
-    "@docusaurus/plugin-ideal-image": "^2.4.3",
-    "@docusaurus/preset-classic": "2.4.3",
-    "@docusaurus/remark-plugin-npm2yarn": "2.4.3",
-    "@docusaurus/theme-live-codeblock": "2.4.3",
-    "@docusaurus/theme-mermaid": "2.4.3",
-    "@mdx-js/react": "^1.6.22",
-    "clsx": "^1.2.1",
+    "@docusaurus/core": "^3.0.1",
+    "@docusaurus/plugin-ideal-image": "^3.0.1",
+    "@docusaurus/preset-classic": "^3.0.1",
+    "@docusaurus/remark-plugin-npm2yarn": "^3.0.1",
+    "@docusaurus/theme-live-codeblock": "^3.0.1",
+    "@docusaurus/theme-mermaid": "^3.0.1",
+    "@mdx-js/react": "^3.0.0",
+    "clsx": "^2.0.0",
     "docusaurus-json-schema-plugin": "^1.7.0",
     "hast-util-is-element": "1.1.0",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "rehype-katex": "5",
-    "remark-math": "3",
+    "prism-react-renderer": "^2.3.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "rehype-katex": "7",
+    "remark-math": "6",
     "swagger-ui-react": "4.15.5"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "2.4.3",
-    "@tsconfig/docusaurus": "^1.0.5",
+    "@docusaurus/module-type-aliases": "3.0.1",
+    "@docusaurus/tsconfig": "3.0.1",
+    "@docusaurus/types": "3.0.1",
+    "@types/react": "^18.0.0",
     "docusaurus-plugin-typedoc": "^0.21.0",
     "dotenv": "^16.3.1",
     "prettier": "^3.0.3",
     "typedoc": "^0.25.3",
     "typedoc-plugin-markdown": "^3.17.0",
-    "typescript": "^4.9.5"
+    "typescript": "~5.2.2"
   },
   "resolutions": {
     "types-ramda": "0.29.4"
